fix: Norm on hooks
